Don't indent HTML and escape strings by default
This commit actually implement a lighter version of printf(), with some
tweaks.  First, strings provided with "%s" are escaped.  This is mostly
what we always want.  One should use "%S" to print a string without any
post processing.

We could have kept indentation by default, and it may be re-implemented
in a later commit.  However I'll keep it simple for now.

One last notable change is that "%d" for intergers don't work anymore,
and "%i" should be used instead.  "%I" expect a long, "%u" an unsigned
and eventually "%U" a long unsigned.

Last but not least, this commit make the sue of html() and it's
derivatives less restrictive at usage.  Indeed, one would had to take
care of starting a format string with "<" and en it with "/>", otherwise
indentation would not work.  This is now gone and *any* format string
are now acceptable.