Change route logic to avoid special case for root route
When the URL was empty (like "/"), we used to have a special case to
make the routing algorithm working.  This commit got ride of this
special case, and for that, changed how default page are found within
the route tree.  Where we had:

	DIR("foo") ... END()
	JSON("foo", foo)
	HTML("foo", foo)

We now have:

	DIR("foo")
		JSON(NULL, foo)
		HTML(NULL, foo)
		...
	END()

And the "joker" page now use an underscore symbol (_) instead of the
string "*".