Introduce json() to properly escape json strings
It's working exactly like html(), but have two additional conversion
specifiers: "%b" for booleans and "%d" for dates.  It also adds quotes
around escaped strings by default, to avoid redundant \" in format
strings.

I ran accross an issue where va_list cannot be passed as a pointer in
functions because va_list can be an array.  I had to encasulate va_list
in a structure to do so.  Every json page has been converted to use the
new json() function, and every remaining functrions in json.[ch] have
been removed, and so are json.[ch] files.

The format of JSON output may have changed.  Player historic format have
changed, it's not really what I want for now, because dates seems messed
up, but it will be fixed in the futur.