Use setjmp()/longjmp() for error handling when generating pages
When a page encountered an error, it had to gracefully unwind the stack
and return the error code.  This leads to a fuckton of untested error
paths, while every errors are handled the same once stack is unwinded.

We set a jump point before routing the request, and just print an error
page when longjmp() is used.  This commit preserve the error() function,
but generalize its used almost everywhere in the CGI.

This change results in a massive simplification of the code of pages.
It will also ease the transition to fast CGI.

There is still a gotcha with pipes: if error() is called after stdout
has been replaced by a pipe, we don't yet restore stdout before printing
the error page.  The error page will then be printed to the pipe that
will not going to be flushed.