Use a memory buffer instead of a pipe to buffer output of pages
The current code made it easy to switch from a pipe to a memory buffer,
because the current code were actually already using a memory buffer,
wich was then written to stdout.

This commit also get add some error() calls to simplify furthermore the
code.  Last but not least it make sure that everything now use
html()-like functions to write data (only "robots.txt" was not).