Move html()-like function in a new file "io.[ch]"
Having them in html.h was inconsistent, so we had to move them.  I
wanted to put them in cgi.c but the file was growing too much: 760
lines.  So choosed to put the code in a new file to not mixing up
unrelated code.

However I include io.h in cgi.h so that they are always available.  And
thus, this commit also removes now useless #include "html.h".