URLs: Don't add paramaters when they have the default value
This changes sound small but is actually quite big and is required to
make the whole website function properly.  First, it allow the code to
make the difference between non-existant parameters (NULL) from existant
but empty ones (empty string "").

The URL() function has been completely reworked to offer this
functionality: we don't use a format string anymore and instead use a
NULL-terminated list of parameters description.  We also store the
result in a buffer because using a statically allocated buffer approach
had issues.  It now looks like this:

	URL(buf, "/foo/bar", PARAM_GAMETYPE("CTF"));

Each parameter have its associated macro PARAM_XXX() for convenience.
This macro give to the URL() function de default value and the type of
the parameter.

This commit also clean up some function the were not using URL(), and
also changed a little bit how dynbuf functions internally.  I have to
admit that C is not really an adequate language for handling strings in
an easy way.  Anyway, now that is finished, I can hopefully move
forward.