Provide a generic way to create HTML/JSON lists
Listing every results from a query requite the exact same logic.  This
commit aim to factorize all this logic into a single set of reusable
functions.

Those functions (one for HTML and one for JSON) take a description fo
each columns of the list and a pointer to the instanciated query.  Then
they generate an output for each row in the result.

Since a lot fo code is actually listing things up, and I fixed some
corner cases along the way.  List should be way more consistent and
easily adjustable.