Replace strdup() by stack allocated buffers
We want to avoid dynamic allocations as much as possible so that we
don't have to worry about memory leaks when we're gonna switch to
fastCGI.

The actual code may look a little ugly, because I need a stack allocated
buffer and a way to handle NULL pointers.  SO I made a macro to define
the pointer as well as the buffer.  I then introduce a function to read
a column from a result row and handle NULL value accordingly.