Make foreach_row() simpler and abstract value extraction from rows
The callack mechanism was kinda annoying and was hard to work with
most of the time.  Plus it expected the user to define
"sqlite3_stmt *res; unsigned nrow;" while not making them explicit in
the argument list.

The new foreach_row() doesn't require unsigned nrow to be defined and
takes a pointer to "sqlite3_stmt" as an argument, making everything
explicit.  It also get ride of the callback mechanism.

Last but not least, casting results of sqlite_column_*() was cumbersome
and error prone.  We introduced column_{int,unsigned,time_t,bool,text}()
to have the expected type right of the bat.

Eventually, the new foreach_row() doesn't have a way to detect errors.
This should not be an issue for now because we always ignore errors
anyway.  I plan to have a global variable like "errno" to indicate a
failure.