Display map alongside gametype in top tabs
Also put tabs of non-vanilla gametype on the left to make them distinct
of custom tabs on the right.  Last but not least, also display the type
of a custom tab : "Player" for player, "Clan" for clan...

The underlying DOM structure of top tab bar and bottom tab bar has been
changed to us <ul> tags, but the end result should be the same.

html_header() now takes optional argument to make its use simpler.