Add a struct list to ease list creation
The goal of this commit is to make list simpler to code.  For this
purpose a structure "list" is added and two functions to initialize it.

The main change is that list order is now automatically handled, so we
don't need to craft the queries depending on the order.  Same geos for
the "LIMIT x OFFSET y" in queries: it is automatically added.

Last but not least, this commit introduce an easy way to check for valid
sorting order with the function check_order().