Add a basic test suite
A custom database for tests is created, and then a bunch of CGI
invocation are made with different URLs.  If one of them does not
returns HTTP 200, we fail.

This test suite is something I wanted for a long time: many time a page
broke without even me noticing.  Now the build will simply fail if a
page returns an error instead of regular content.