Fix clan list including empty clan
The empty clan is not really a clan, so we should not include it in any
results.  Also, some queries were trying to filter out the empty clan by
using something like: "clan IS NO NULL".  But NULL is not the empty
string, so this commit also rewrite those queries by "clan <> ''".