list.c: Don't use IFNULL() trick because map cannot be NULL anymore
Indeed: PARAM_MAP have now "" as a default value, wich make sure
EXTRACT_URL() cannot return a NULL value when feeded with PARAM_MAP.

The case for server a little bit different, because we don't want to
filter them by map when map is "".  We previsouly set a NULL value when
"map" was NULL, and used IFNULL() in the query to handle it, but
actually, just testing for "map = '' OR foo = map" is enough.