Handle extended packets and legacy 64 packets
This allow us to handle servers with more than 16 players, like ddrace
servers or city servers.  My code is based on ddnet code, and mimic the
same behavior, with some minor differences.

Even though legacy 64 servers are doomed to disappear, there is still at
least one of them online.  So it was worth supporting this packet
format.

The unpacker had to be reworked as well as the packet handling logic,
because player list can be spread in multiple packets.  I reworked the
unpacker to make unpacking simpler, and also to not print error message
on standard error output when a packet is malformed.