Poll vanilla and legacy64 servers for teeworlds 0.6.
First working version.  It is capable of retrieving the list of all 0.6
servers from the master servers and then polling all servers.  It supports
vanilla and legacy64 format, but not the extended format.  It then proceed to
log what it got.

There is still some errors here and there but at least we now have a working
base.  Also the new major release 0.7 came out some time ago and the protocol
changed, that's why I'm sticking with 0.6 at the moment because I know the
protocol and already used it for a previous version of Teerank.