Drop packets with the wrong request token.
If a packet somehow takes more than a second to arrive, then it will processed
by the next poll and it will likely confuse the processing code because it
will receive twice as more packet than expected.

To be able to craft custom packets with request token, the packet code had to
be changed to be less generic and more customizable.

When sending a request token, any modded servers will likely answer with
a 'iext' packet rather than a 'gie3' packet as we used to get.  We don't
handle 'iext' packets yet.