Raise a PacketException for malformed packets.
This execption is catched when server_pool calls server.process_packet().  Any
malformed pack will be dropped and the reason will be logged.

Because using int() on a python string may raise a ValueError if the string
isnot a parsable integer, we added a packet.unpack_int() which transforms a
ValueError into a PacketException.