Bind backend socket to port 8311.
To query server on internet through a docker container, outgoing ports must be
explicitely declared as so.  If bind() is not used on the socket a random
port is choosed, which is probably not going to be the port declared to
docker.  To solve this bind socket to port 8311.

Why port 8311?  Because most teeworlds servers use ports between 8303-3310, so
it makes 8311 a natural port for Teerank.  Of course this port is not already
used by a well known application.