Fix dangling client reference in game servers.
When replacing a server clients list, clients were deleted but the reference
was kept inside the 'clients' list.  As a result any query about clients
failed because of this dangling reference.

Fixing this required us to put a reference back to the game server for each
client.  That way when deleting the client, DGraph will automatically remove
it from the game server clients list.