Try to avoid given dgraph strings with \u in it, or any other sequence really.
Not sure if it's a Dgraph issue or more of a JSON encoding issue but strings
with escape sequence confuse the dgraph lexer and as a result make queries
fail.

The fix try to use a different fallback method for decoding strings coming
from any packets.  Instead of using 'replace', we switched
to 'backslashreplace', which seems to make things work.  Although I'm still
not 100% confident that this will fix all the issues with dgraph lexer.