Move database modules from shared/ to backend/.
In an attempt to make frontend an indenpendant service, remove any shared
logic between both.  While the dgraph connection code was the only true bit
of code shared by both services, I think it is still valuable to have no
shared code.

Replace the global databases handle by a function that return an handle.  This
done in order to avoid having code run at load time to initialize the
database handle.  Such code is problematic because it has side effects.  For
instance when testing we want to control how and when the database connection
is made to make everything explicit.