Register routes when creating the application.
This remove the need to use a blueprint and also remove a little hack
using __init__.py.

Also I was working on testing each page individually with pytest and blueprint
messed things up to a point where I decided to try to get ride of blueprint
to see if it worked and it did.  So that's another reason to not use
blueprint if we can.