Use the same name for database containers for both -dev and -test.
Previously databases containers needed different name between -dev and -test
because they were sharing the same project name and as such using the same
name meant sharing database for both -dev and -test, which was obviously not
the intended.

However now that we use different project name for -dev and -test, we can
safely use the same database containers name.