Add the basic foundation for ranking per gametype and per map.
First add graphql types GameType and Map.  Each gametype has its own list of
map.  When GameType name is None then it is a special gametype entry for all
gametypes.  The same goes for when map name is None.

Each map then has a PlayerInfo list that store the score for a given player.
This new way of storing players score replace the redis sorted set we used
previously.

The rank() function now adds entries for new players score for the currnet map
as well as for all maps of the current gametype.