Fix Dgraph not handling some escape sequences.
This fix required an extra step before sending data to Dgraph.  It also
erquired adding another step to cancel the effect of the first step for data
received from dgraph.

Since both the frontend and backend needs to do this transformation, the
shared/ folder was added back.  Also to make sure all the code use this
serialization and unserialization, a new way to execute queries have been
introduced and the graphql handle is now hidden.